BizScope/requirements.txt
requests
azure-functions
azure-eventhub